requirements.txt
adal==0.4.7
asn1crypto==0.23.0
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage==0.36.0
azure-storage-blob==0.37.1
azure-storage-common==0.37.1
azure-storage-nspkg==2.0.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
confusable-homoglyphs==2.0.2
cryptography==2.1.3
Django==1.11.6
django-betterforms==1.1.4
django-constrainedfilefield==3.0.7
django-registration==2.3
django-sendgrid-v5==0.6.84
django-storages==1.6.5
enum34==1.1.6
futures==3.1.1
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.18
isodate==0.6.0
jsonschema==2.6.0
keyring==10.4.0
libmagic==1.0
msrest==0.4.18
msrestazure==0.4.16
oauthlib==2.0.6
pathlib2==2.3.0
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.3.0
python-dateutil==2.6.1
python-http-client==3.0.0
python-magic==0.4.13
pytz==2017.2
pywin32-ctypes==0.1.2
requests==2.18.4
requests-oauthlib==0.8.0
scandir==1.6
sendgrid==5.3.0
six==1.11.0
urllib3==1.22
validictory==1.1.1
whitenoise==3.3.1
